requirements.txt
pyautogui
Pillow >= 4.0.0